Big news
- Introducing Newtoo Web Interface Mapper  — the new technology converting WebIDL to C++
https://github.com/FlightBlaze/Newtoo-web-interface-mapper

- Announced Newtoo Styling Engine. See IDL Index at https://github.com/FlightBlaze/Newtoo/blob/master/modules/styling/index.idl

- Double instead of integers in DOM Geometry

- New SVG Interfaces

+++ modules/dom/geometry/DOMRect.cpp
@@ -3,23 +3,23 @@
 namespace Newtoo
 {
 
-    DOMRect::DOMRect(int aX, long aY, int aWidth, long aHeight)
+    DOMRect::DOMRect(double aX, double aY, double aWidth, double aHeight)
         :DOMRectReadOnly(aX, aY, aWidth, aHeight)
     {}
 
-    void DOMRect::setX(int aX)
+    void DOMRect::setX(double aX)
     {
         mX = aX;
     }
-    void DOMRect::setY(int aY)
+    void DOMRect::setY(double aY)
     {
         mY = aY;
     }
-    void DOMRect::setWidth(int aWidth)
+    void DOMRect::setWidth(double aWidth)
     {
         mWidth = aWidth;
     }
-    void DOMRect::setHeight(int aHeight)
+    void DOMRect::setHeight(double aHeight)
     {
         mHeight = aHeight;
     }
